requirements.txt
psutil
discord.py>=2.3.0
yt-dlp>=2024.0.0
aiohttp>=3.9.0
pillow>=10.0.0
python-dotenv>=1.0.0
lyricsgenius>=3.0.0
gTTS>=2.5.0
aiodns>=3.0.0
certifi
PyNaCl>=1.5.0
google-generativeai>=0.3.0
jikanpy>=4.3.0
quart>=0.20.0